DEBUG: Implement some heinous hacks to allow testing SQL interface
Make 'DESCRIBE ALTER <string literal>' interpret string literal as
raw SQL source, run it against the database (by piping it to psql!)
and return the textual result.

I went with DESCRIBE ALTER as the (totally nonsensical) keyword for
this because I didn't want to have to adjust any keywords, it worked,
and the presence of ALTER prevents the query normalizer from
extracting the string constant.